+++ fakernaija/tests/test_schools.py
@@ -62,7 +62,7 @@ class TestSchoolProvider(unittest.TestCase):
     def test_load_json_file_not_found(self, mock_file: Any) -> None:  # noqa: ANN401
         """Test handling of file not found error."""
         mock_file.side_effect = FileNotFoundError
-        with self.assertRaises(FileNotFoundError):  # noqa: PT027
+        with self.assertRaises(FileNotFoundError):
             self.school_provider.load_json("none.json")
 
     @patch(
@@ -75,7 +75,7 @@ class TestSchoolProvider(unittest.TestCase):
         mock_file: Any,  # noqa: ARG002, ANN401
     ) -> None:
         """Test handling of invalid JSON format."""
-        with self.assertRaises(ValueError) as context:  # noqa: PT027
+        with self.assertRaises(ValueError) as context:
             self.school_provider.load_json(self.school_provider.data_path)
         assert "Error decoding JSON" in str(context.exception)
 
@@ -89,7 +89,7 @@ class TestSchoolProvider(unittest.TestCase):
         mock_file: Any,  # noqa: ARG002, ANN401
     ) -> None:
         """Test handling of invalid data structure."""
-        with self.assertRaises(ValueError) as context:  # noqa: PT027
+        with self.assertRaises(ValueError) as context:
             self.school_provider.load_json(self.school_provider.data_path)
         assert "Invalid data format" in str(context.exception)
 
@@ -103,7 +103,7 @@ class TestSchoolProvider(unittest.TestCase):
         mock_file: Any,  # noqa: ARG002, ANN401
     ) -> None:
         """Test handling of missing required keys in school data."""
-        with self.assertRaises(ValueError) as context:  # noqa: PT027
+        with self.assertRaises(ValueError) as context:
             self.school_provider.load_json(self.school_provider.data_path)
         assert "Invalid data format" in str(context.exception)